Update types in drawing function

+++ src/interpreter.cc
@@ -177,18 +177,18 @@ void Interpreter::OP_Cxkk() {
 
 // Display n-byte sprite starting at memory location I at (Vx, Vy), set VF = collision
 void Interpreter::OP_Dxyn() {
-    unsigned short Vx = c8.registers[getX()];
-    unsigned short Vy = c8.registers[getY()];
-    unsigned short height = c8.opcode & 0x000F;
-    unsigned short pixel;
+    uint16_t Vx = c8.registers[getX()];
+    uint16_t Vy = c8.registers[getY()];
+    uint16_t height = c8.opcode & 0x000F;
+    uint16_t pixel;
 
     c8.registers[0xF] = 0;
 
-    for (int y = 0; y < height; y++) {
+    for (size_t y = 0; y < height; y++) {
         pixel = c8.memory[c8.I + y];
-        for (int x = 0; x < 8; x++) {
+        for (size_t x = 0; x < 8; x++) {
             if (pixel & (0x80u >> x)) {
-                unsigned int loc = x + Vx + (y + Vy) * c8.display.width();
+                uint32_t loc = x + Vx + (y + Vy) * c8.display.width();
                 if (c8.display[loc]) {
                     c8.registers[0xF] = 1;
                 }
